- some clean-ups:     word -> uint16     new definitions INT16_MIN/MAX and INT32_MIN/MAX replacing SHORT_MIN/MAX       and LONG_MIN/MAX in cstypes.h     SHORT_MIN/MAX -> INT16_MIN/MAX or SHRT_MIN/MAX from limits.h, as       appropriate     boolean -> bool     TRUE/FALSE -> true/false - integrated FreeBSD patches from Michael Alyn Miller - OpenedResourceFile::Check()/Get() can now also take 4 chars instead of a   uint32 resource type - wadfile checksums are now uint32 consistently, instead of "long" or   "unsigned long"
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@393 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Pfhortran/script_parser.h
@@ -4,6 +4,8 @@
 #ifndef _SCRIPT_PARSER_DEF
 #define _SCRIPT_PARSER_DEF
 
+#include "cstypes.h"
+
 enum /* procedure traps */
 {
 	idle = 1,
@@ -27,7 +29,7 @@ struct bind_table
 	bool active;
 	int which;
 	bool available;
-	long instruction_decay;
+	uint32 instruction_decay;
 };
 
 struct script_instruction
@@ -57,7 +59,7 @@ int get_trap_start(short trap);
 void set_trap_instruction(short trap, int offset);
 void reset_trap(short trap);
 void activate_trap(short trap);
-long get_trap_instruction_decay(short trap);
-void set_trap_instruction_decay(short trap, long decay);
+uint32 get_trap_instruction_decay(short trap);
+void set_trap_instruction_decay(short trap, uint32 decay);
 
 #endif
